refactored http response parser
